Reuse caller optimisation in functions
Method and function calls are very similar in nature but used different code paths: method calls have an optimisation where an optimised caller trampoline is generated once and reused for each function signature, whereas regular call serialized arguments as type-erased values each time.

There is no reason for them to diverge, so I made function calls reuse same optimisation.

Results on a silly microbenchmark running a side-effect-ful JS function with 100 integer args in a loop:

as method: 224.504ns
as function: 5367.9ns

as method: 845.044ns
as function: 5350.96ns

as method: 262.74ns
as function: 246.593ns

as method: 859.534ns
as function: 823.498ns

In real world the difference is unlikely to be perceptible, but at the very least this makes the code more consistent.