Convert run_multiple_processes to asyncio
When running lots of processes, I noticed in a profiler that a lot of time is spent in small chunks doing nothing, just waiting for a free process.

After looking at the code, I figured why - it was waiting up to 0.2ms for each process in a loop until it found a free one, which could add up.

This kind of tasks is exactly what asyncio is designed to solve - it can wait for subprocesses concurrently and return a free one as soon as it's available.